Fix drift
Previously the flock would always drift
up and to the left. This was fixed by
changing the bird tick function to take the
abs() of x and y separetly instead of the abs()
of their difference.

The keep on screen function steals some of the
"total energy" of the flock when a bird tries to
leave the screen. Not ideal. Need some way to
add "energy" to the system.